+++ src/application/gleis/gerade.rs
@@ -5,7 +5,7 @@ use std::{fmt::Debug, hash::Hash, marker::PhantomData};
 use serde::{Deserialize, Serialize};
 use zugkontrolle_derive::alias_serialisiert_unit;
 
-use super::anchor;
+use super::verbindung;
 use crate::anschluss::{InputAnschluss, InputSerialisiert};
 use crate::steuerung::kontakt::Kontakt;
 use crate::{application::typen::*, lookup::impl_lookup};
@@ -39,7 +39,7 @@ impl<Z> GeradeUnit<Z> {
     }
 }
 
-#[impl_lookup(anchor::Anchor, Points)]
+#[impl_lookup(verbindung::Anchor, Points)]
 #[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
 pub enum AnchorName {
     Anfang,
@@ -85,11 +85,11 @@ impl<Z: Zugtyp, Anschluss: MitName> Zeichnen for Gerade<Z, Anschluss> {
         let gleis_rechts = gleis_links + self.länge;
         let beschränkung_mitte = beschränkung::<Z>().halbiert();
         AnchorPoints {
-            anfang: anchor::Anchor {
+            anfang: verbindung::Anchor {
                 position: Vektor { x: gleis_links, y: beschränkung_mitte },
                 richtung: winkel::PI,
             },
-            ende: anchor::Anchor {
+            ende: verbindung::Anchor {
                 position: Vektor { x: gleis_rechts, y: beschränkung_mitte },
                 richtung: winkel::ZERO,
             },