+++ src/application/gleis/verbindung.rs
@@ -1,11 +1,12 @@
-//! anchor points to mark connection points of a rail
+//! Verbindung zwischen zwei Gleisen
 
 use crate::lookup;
 
 pub mod point;
-pub(crate) mod rstar;
+pub(crate) mod rstern;
 
 pub use point::*;
 
+/// Spezialisierung des Lookup-Traits auf Verbindung als Element.
 pub trait Lookup<Name>: lookup::Lookup<Name, Anchor> {}
 impl<Name, T: lookup::Lookup<Name, Anchor>> Lookup<Name> for T {}